fix: Handle pandas 2.x compatibility for numpy integer types and timestamps
Use np.integer (base class) instead of np.int_ (int64 only) since pandas
2.x groupby can return np.int32 keys. Switch test helpers from
.values.tolist() to .tolist() so timestamps remain as datetime objects
instead of being converted to raw int64 nanoseconds.


rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED

+++ caikit/interfaces/ts/data_model/backends/pandas_backends.py
@@ -206,7 +206,7 @@ class PandasTimeSeriesBackend(TimeSeriesBackendBase):
         error.type_check_all(
             "<COR81128383E>",
             str,
-            np.int_,
+            np.integer,
             int,
             allow_none=True,
             ids=ids,
@@ -255,10 +255,10 @@ class PandasTimeSeriesBackend(TimeSeriesBackendBase):
             return self._timestamp_column
 
         if name == "ids" and self._ids is not None and len(self._ids) != 0:
-            if isinstance(self._ids[0], (np.int_, int)):
+            if isinstance(self._ids[0], (np.integer, int)):
                 val = data_model_class.IntIDSequence(
                     values=[
-                        id.item() if isinstance(id, np.int_) else id for id in self._ids
+                        id.item() if isinstance(id, np.integer) else id for id in self._ids
                     ]
                 )
                 return DataBase.OneofFieldVal(val=val, which_oneof="id_int")
